Streaming: Refactor redis connection login to separate module
